# packages/contracts/openapi/dabboba.openapi.yaml
openapi: 3.1.0
info:
  title: DABBOBA API
  version: 0.1.0
  description: Shared user, commerce, and privileged operations contract. Fixture-only prototype behavior is not represented as production persistence.
servers:
  - url: http://127.0.0.1:8788
    description: Isolated local API
...
paths:
...
  /v1/wanted-requests:
    get:
      tags: [Community, Catalog]
      operationId: listWantedRequests
      description: Public active wanted-request room. Supplying a valid user session adds the viewer's like state.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
        - name: category
          in: query
          schema: { $ref: "#/components/schemas/ProductCategory" }
        - name: ipId
          in: query
          schema:
            type: string
            maxLength: 120
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
      responses:
        "200":
          description: Active wanted requests ordered newest first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WantedRequestPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Community, Catalog]
      operationId: createWantedRequest
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWantedRequestInput" }
      responses:
        "201":
          description: Wanted request created with a typed work name, an optional catalog IP match, and an optional photo.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WantedRequest" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/wanted-requests/{requestId}:
    patch:
      tags: [Community, Catalog]
      operationId: updateWantedRequest
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/WantedRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateWantedRequestInput" }
      responses:
        "200":
          description: Author-owned active wanted request updated after a version check.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WantedRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Community, Catalog]
      operationId: deleteWantedRequest
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/WantedRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExpectedVersionInput" }
      responses:
        "200":
          description: Author-owned wanted request soft-deleted after a version check.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WantedRequestDeletionResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/wanted-requests/{requestId}/like:
    post:
      tags: [Community]
      operationId: setWantedRequestLike
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/WantedRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetWantedRequestLikeInput" }
      responses:
        "200":
          description: Explicit viewer like state and committed aggregate count.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WantedRequestLikeResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 500 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
    Search:
      name: q
      in: query
      schema: { type: string, maxLength: 120 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 200 }
...
    WantedRequestId:
      name: requestId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Request input is malformed or references an inactive catalog item.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Authentication is required or the session is invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: The authenticated actor lacks the required role or ownership.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: The visible resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: The requested state transition conflicts with committed state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }
...
    ContentStatus: { type: string, enum: [ACTIVE, HIDDEN, DELETED] }
...
    ProductCategory: { type: string, enum: [gacha, figure, kuji, tcg] }
...
    PageMeta:
      type: object
      additionalProperties: false
      required: [nextCursor]
      properties:
        nextCursor: { type: [string, "null"] }
...
    ExpectedVersionInput:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, minimum: 1 }
...
    WantedRequest:
      type: object
      additionalProperties: false
      required: [id, userId, authorNickname, category, ipId, ipNameKo, desiredItem, details, mediaId, status, likeCount, likedByViewer, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        authorNickname: { type: string }
        category: { $ref: "#/components/schemas/ProductCategory" }
        ipId:
          type: [string, "null"]
          maxLength: 120
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        ipNameKo: { type: string, minLength: 1, maxLength: 160 }
        desiredItem: { type: string, minLength: 1, maxLength: 240 }
        details: { type: string, minLength: 1, maxLength: 5000 }
        mediaId: { type: [string, "null"], format: uuid }
        status: { $ref: "#/components/schemas/ContentStatus" }
        likeCount: { type: integer, minimum: 0 }
        likedByViewer: { type: boolean }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateWantedRequestInput:
      type: object
      additionalProperties: false
      required: [category, ipNameKo, desiredItem, details]
      properties:
        category: { $ref: "#/components/schemas/ProductCategory" }
        ipId:
          type: [string, "null"]
          maxLength: 120
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        ipNameKo: { type: string, minLength: 1, maxLength: 160 }
        desiredItem: { type: string, minLength: 1, maxLength: 240 }
        details: { type: string, minLength: 1, maxLength: 5000 }
        mediaId: { type: [string, "null"], format: uuid }
    SetWantedRequestLikeInput:
      type: object
      additionalProperties: false
      required: [liked]
      properties:
        liked: { type: boolean }
    WantedRequestLikeResult:
      type: object
      additionalProperties: false
      required: [requestId, liked, likeCount]
      properties:
        requestId: { type: string, format: uuid }
        liked: { type: boolean }
        likeCount: { type: integer, minimum: 0 }
    WantedRequestPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/WantedRequest" } }
...
    UpdateWantedRequestInput:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, minimum: 1 }
        category: { $ref: "#/components/schemas/ProductCategory" }
        ipId:
          type: [string, "null"]
          maxLength: 120
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        ipNameKo: { type: string, minLength: 1, maxLength: 160 }
        desiredItem: { type: string, minLength: 1, maxLength: 240 }
        details: { type: string, minLength: 1, maxLength: 5000 }
        mediaId: { type: [string, "null"], format: uuid }
      description: At least one mutable property in addition to expectedVersion is required.
    WantedRequestDeletionResult:
      type: object
      additionalProperties: false
      required: [requestId, status, version]
      properties:
        requestId: { type: string, format: uuid }
        status: { type: string, const: DELETED }
        version: { type: integer, minimum: 1 }
...